pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fusion-reviewer"
version = "0.1.0"
description = "中文优先的委员会式论文审稿系统，支持 PDF / DOCX / DOC、MinerU 预处理、Codex skill 与本地 backend 双模式。"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
  "click>=8.1.7",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "openai>=2.26,<3",
  "anthropic>=0.34.0",
  "google-generativeai>=0.8.0",
  "pypdf>=4.3.0",
  "pymupdf>=1.24.9",
  "reportlab>=4.2.0",
  "brotli>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
fusion-review = "fusion_reviewer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--basetemp=.tmp/pytest-basetemp"
cache_dir = ".tmp/pytest-cache"
